backend/src/requirements_docker.txt
ariadne==0.21
fastapi==0.109.1
httpx==0.26.0
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.7
python-jose==3.3.0
SQLAlchemy==2.0.24
uvicorn==0.25.0

